feat!: add mandatory environment variable for service id
Adds a new `DSP_SERVICE_ID` environment variable through which to set the
identifier of the service. This is the identifier with which the service itself
is uniquely identified in the dataspace.

BREAKING CHANGE: The new `DSP_SERVICE_ID` environment variable is mandatory and
the service will not work without it being set.